# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QA Pulse - QA Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # Requirements
  /requirements:
    get:
      operationId: listRequirements
      tags: [requirements]
      summary: List requirements
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: module
          in: query
          schema:
            type: string
        - name: release
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requirement"
    post:
      operationId: createRequirement
      tags: [requirements]
      summary: Create requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequirementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
  /requirements/{id}:
    get:
      operationId: getRequirement
      tags: [requirements]
      summary: Get requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
    patch:
      operationId: updateRequirement
      tags: [requirements]
      summary: Update requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequirementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
    delete:
      operationId: deleteRequirement
      tags: [requirements]
      summary: Delete requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Test Cases
...
  /social-events:
    get:
      operationId: listSocialEvents
      tags: [social]
      summary: List social/team hangout events
      responses:
        "200":
          description: List of social events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialEvent"
    post:
      operationId: createSocialEvent
      tags: [social]
      summary: Create a social/team hangout event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialEvent"
  /social-events/{id}:
    patch:
      operationId: updateSocialEvent
      tags: [social]
      summary: Update a social event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialEvent"
    delete:
      operationId: deleteSocialEvent
      tags: [social]
      summary: Delete a social event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string

    # Requirement
    Requirement:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        module:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        priority:
          type: string
        release:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        redmineTicketId:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    RequirementInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        module:
          type: string
        projectId:
          type: integer
        priority:
          type: string
        release:
          type: string
        assigneeId:
          type: integer
        redmineTicketId:
          type: string
        status:
          type: string
    RequirementUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        module:
          type: string
        projectId:
          type: integer
        priority:
          type: string
        release:
          type: string
        assigneeId:
          type: integer
        redmineTicketId:
          type: string
        status:
          type: string

    # Test Case
...
    SocialEvent:
      type: object
      required: [id, title, eventDate, eventType, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        eventDate:
          type: string
          description: "YYYY-MM-DD"
        eventType:
          type: string
          description: "lunch | dinner | birthday | outing | other"
        taggedUserIds:
          type: array
          items:
            type: integer
        taggedUserNames:
          type: array
          items:
            type: string
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
    SocialEventInput:
      type: object
      required: [title, eventDate, eventType]
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        eventType:
          type: string
        taggedUserIds:
          type: array
          items:
            type: integer
        createdBy:
          type: integer
    SocialEventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        eventType:
          type: string
        taggedUserIds:
          type: array
          items:
            type: integer