# lib/api-zod/src/generated/types/requirement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Pulse - QA Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Requirement {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  module?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  priority: string;
  /** @nullable */
  release?: string | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  redmineTicketId?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}